1.9.0-beta
Changelog
- Se han corregido errores menores:
1.-Se ha corregido el salto de linea en el login al introducir el usuario y contraseña.
2.-El campo solicitaba el email cuando se autenticaba con el usuario, se ha corregido el texto, ahora solicita correctamente el nombre de usuario.
3.-Se ha corregido un error que no permitía la introducción de decimales en la puntuación de los platos.
4.- Se ha corregido un bug que permitía que el filtro de solicitudes en la api pudiera falla
-Nuevas Funciones:
1.- Se ha añadido un buscador de restaurante a la hora de asignar el restaurante en un nuevo plato
2.- Se ha añadido una opción solo para administradores para asignar roles a diferentes  usuarios
3.- Se ha añadido la opción de editar los platos
4.- Se ha añadido solo para administradores la opción de editar los restaurantes.

+++ LoveThings/LoveThings/src/main/java/core/neoarcadia/lovethings/models/Dish.java
@@ -18,9 +18,9 @@ public class Dish {
 
     private Double price;
 
-    private Integer waitTime;
+    private Double waitTime;
 
-    private Integer rating;
+    private Double rating;
 
     private String notes;
     private Boolean isFavorite = false;
@@ -66,19 +66,19 @@ public class Dish {
         this.price = price;
     }
 
-    public Integer getWaitTime() {
+    public Double getWaitTime() {
         return waitTime;
     }
 
-    public void setWaitTime(Integer waitTime) {
+    public void setWaitTime(Double waitTime) {
         this.waitTime = waitTime;
     }
 
-    public Integer getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
